+++ index.d.ts
@@ -5,13 +5,13 @@
  * @param {string[]} [locales]
  */
 export function loadI18NResource(namespace: string, dirResource: string, locales?: string[]): void;
-export function T(key: import('i18next').TFuncKey, options: import('i18next').TOptions, lng: string): any;
+export function T(key: import('i18next').TFuncKey, options: import('i18next').TOptions, locale: string): string;
 /**
  * @param {string} namespace
  * @param {string[]} [locales]
  * @param {string[]} [formats]
- * @returns {TranslatorLocale}
+ * @returns {TranslatorWithGlobalLocale}
  */
-export function TT(namespace: string, locales?: string[], formats?: string[]): TranslatorLocale;
-export type Translator = (key: import('i18next').TFuncKey, options: import('i18next').TOptions, lng: string) => any;
-export type TranslatorLocale = (key: import('i18next').TFuncKey, options: import('i18next').TOptions) => any;
+export function TT(namespace: string, locales?: string[], formats?: string[]): TranslatorWithGlobalLocale;
+export type TranslatorWithLocale = (key: import('i18next').TFuncKey, options: import('i18next').TOptions, locale: string) => string;
+export type TranslatorWithGlobalLocale = (key: import('i18next').TFuncKey, options: import('i18next').TOptions) => string;